fixed background component
